+++ pkg/azure/tokencredentials/ucpcredentials_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	sdk_cred "github.com/project-radius/radius/pkg/sdk/credentials"
+	sdk_cred "github.com/project-radius/radius/pkg/ucp/credentials"
 )
 
 type mockProvider struct {
