+++ pkg/ucp/aws/ucpcredentialprovider_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"helm.sh/helm/v3/pkg/time"
 
-	sdk_cred "github.com/project-radius/radius/pkg/sdk/credentials"
+	sdk_cred "github.com/project-radius/radius/pkg/ucp/credentials"
 )
 
 type mockProvider struct {
