+++ pkg/ucp/frontend/api/server.go
@@ -17,8 +17,8 @@ import (
 	"github.com/project-radius/radius/pkg/armrpc/servicecontext"
 	"github.com/project-radius/radius/pkg/middleware"
 	"github.com/project-radius/radius/pkg/sdk"
-	sdk_cred "github.com/project-radius/radius/pkg/sdk/credentials"
 	ucpaws "github.com/project-radius/radius/pkg/ucp/aws"
+	sdk_cred "github.com/project-radius/radius/pkg/ucp/credentials"
 	"github.com/project-radius/radius/pkg/ucp/dataprovider"
 	"github.com/project-radius/radius/pkg/ucp/frontend/controller"
 	ctrl "github.com/project-radius/radius/pkg/ucp/frontend/controller"
